TS-1252: stats summary in cluster not working
Here is some stats with evil result, we all know that most stats
with proxy.process.cluster prefix have no result at all.

   proxy.process.cluster.open_delays=-4411607530454032024
   proxy.process.cluster.control_messages_avg_send_time=-72.902885
   proxy.process.cluster.open_delay_time=-2494.564697
   proxy.process.cluster.rmt_cache_callback_time=-747.036865
   proxy.process.cluster.local_connection_time=-15.002351

When int64_t integer was passed to CLUSTER_SUM_DYN_STAT(x, y) as
2th parameter, it would be convert to int unsafely.

+++ iocore/cluster/P_Cluster.h
@@ -131,7 +131,7 @@ extern RecRawStatBlock *cluster_rsb;
 #define CLUSTER_DECREMENT_DYN_STAT(x) \
 	RecIncrRawStat(cluster_rsb, mutex->thread_holding, (int) x, -1);
 #define CLUSTER_SUM_DYN_STAT(x, y) \
-	RecIncrRawStat(cluster_rsb, mutex->thread_holding, (int) x, (int) y);
+	RecIncrRawStat(cluster_rsb, mutex->thread_holding, (int) x, y);
 #define CLUSTER_SUM_GLOBAL_DYN_STAT(x, y) \
 	RecIncrGlobalRawStatSum(cluster_rsb,x,y)
 #define CLUSTER_CLEAR_DYN_STAT(x) \
